Move mocks test
