+++ src/__tests__/api/firebase/user.jest.test.ts
@@ -1,10 +1,10 @@
-import "../../../__mocks__/apiMockNext13"
-import "../../../__mocks__/uuidMock"
-import "../../../__mocks__/firebaseAdminMock"
-import { setEnv } from "../../../__mocks__/setEnv"
+import "@/__tests__/mocks/apiMockNext13"
+import "@/__tests__/mocks/uuidMock"
+import "@/__tests__/mocks/firebaseAdminMock"
+import { setEnv } from "@/__tests__/mocks/setEnv"
 import { POST, GET } from "@/app/api/firebase/user/route"
 import { NextRequest } from "next/server"
-import { sampleFirebaseConfig } from "../../../__mocks__/firebaseAdminMock"
+import { sampleFirebaseConfig } from "@/__tests__/mocks/firebaseAdminMock"
 
 describe("firebase/user", () => {
   const mockRequest = (json = {}) => {
