+++ src/__tests__/mocks/snakeWasmMock.ts
