+++ src/__tests__/mocks/styleMock.js
