+++ src/components/Chat/ChatMessageBox/ChatMessageDialog/CustomMessageRender/CustomMessageRender.jest.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import CustomMessageRender from "."
-import "../../../../../__mocks__/windowMock"
+import "@/__tests__/mocks/windowMock"
 
 describe("CustomMessageRender", () => {
   it("should render the component message correctly", () => {
