+++ src/components/CommandBar/CommandBar.jest.test.tsx
@@ -1,4 +1,4 @@
-import "../../__mocks__/routerMock"
+import "@/__tests__/mocks/routerMock"
 import { render, screen } from "@testing-library/react"
 import CommandBar from "."
